🔥 Remove ironic disposing
The disposing as implemented before, using the finaliser would never work, since there is no GL context associated with the finaliser thread. Hence we might as well stick to simpler Dispose methods and rely on callers to be on the right thread (just like before) until we find a better solution.