🔥 Remove IVertexData.Size
This has not been needed for a while. The only place we care about the size (VertexBuffer) already uses the VertexData.Size directly. This is probably slightly slower than caching it in the actual type, but we could add our own cache here in the future if this should ever turn into a bottleneck, which is unlikely.